Complete draft of pullprofiles
Made script executable

Added formatedate function to add to mysql database

Created gettags iterator to access tags in database. I'm experementing
with a return statement to raise a StopIteration exception. If this
fails, then I might have to replce the line with a self.close() call to
cleanly close the iterator.

Added exception in the mapentries function to catch general errors. I'm
not sure if this is what I want, as the blogentries may pass commas that
raise strange errors.